backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.27.0
Pillow>=11.0.0
markdown>=3.6
pypdf>=4.2.0
python-docx>=1.1.0
google-generativeai>=0.8.0
groq>=0.5.0
mistralai>=1.0.0
